+++ baseapp/coin.go
@@ -1,4 +1,4 @@
-package types
+package baseapp
 
 import (
 	"fmt"
