+++ baseapp/errors.go
@@ -3,171 +3,96 @@ package baseapp
 import (
 	"fmt"
 
-	abci "github.com/tendermint/abci/types"
-	cmn "github.com/tendermint/tmlibs/common"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
-// ABCICodeType - combined codetype / codespace
-type ABCICodeType uint32
-
-// CodeType - code identifier within codespace
-type CodeType uint16
-
-// CodespaceType - codespace identifier
-type CodespaceType uint16
-
-// root error codes
+// SDK error codes
 const (
-
 	// ABCI error codes
-	ABCICodeOK ABCICodeType = 0
+	ABCICodeOK sdk.ABCICodeType = 0
 
-	// Root error codes
-	CodeOK             CodeType = 0
-	CodeInternal       CodeType = 1
-	CodeUnknownRequest CodeType = 2
-	CodeTxDecode       CodeType = 3
+	// Base error codes
+	CodeOK                sdk.CodeType = 0
+	CodeInternal          sdk.CodeType = 1
+	CodeTxDecode          sdk.CodeType = 2
+	CodeInvalidSequence   sdk.CodeType = 3
+	CodeUnauthorized      sdk.CodeType = 4
+	CodeInsufficientFunds sdk.CodeType = 5
+	CodeUnknownRequest    sdk.CodeType = 6
+	CodeInvalidAddress    sdk.CodeType = 7
+	CodeInvalidPubKey     sdk.CodeType = 8
+	CodeUnknownAddress    sdk.CodeType = 9
+	CodeInsufficientCoins sdk.CodeType = 10
+	CodeInvalidCoins      sdk.CodeType = 11
 
-	// CodespaceRoot is a codespace for error codes in this file only.
-	// Notice that 0 is an "unset" codespace, which can be overridden with
-	// Error.WithDefaultCodespace().
-	CodespaceUndefined CodespaceType = 0
-	CodespaceRoot      CodespaceType = 1
-
-	// Maximum reservable codespace (2^16 - 1)
-	MaximumCodespace CodespaceType = 65535
+	CodespaceSDK sdk.CodespaceType = 2
 )
 
-// Error type
-type Error interface {
-	Error() string
-	Code() CodeType
-	Codespace() CodespaceType
-	ABCILog() string
-	ABCICode() ABCICodeType
-	WithDefaultCodespace(codespace CodespaceType) Error
-	Trace(msg string) Error
-	T() interface{}
-	Result() Result
-	QueryResult() abci.ResponseQuery
-}
-
-// get the abci code from the local code and codespace
-func ToABCICode(space CodespaceType, code CodeType) ABCICodeType {
-	// TODO: Make Tendermint more aware of codespaces.
-	if space == CodespaceRoot && code == CodeOK {
-		return ABCICodeOK
-	}
-	return ABCICodeType((uint32(space) << 16) | uint32(code))
-}
-
-// IsOK - is everything okay?
-func (code ABCICodeType) IsOK() bool {
-	if code == ABCICodeOK {
-		return true
-	}
-	return false
-}
-
-// NewError - create an error
-func NewError(codespace CodespaceType, code CodeType, msg string) Error {
-	return &StdError{
-		codespace: codespace,
-		code:      code,
-		err:       cmn.NewErrorWithT(code, msg),
+// NOTE: Don't stringer this, we'll put better messages in later.
+func CodeToDefaultMsg(code sdk.CodeType) string {
+	switch code {
+	case CodeInternal:
+		return "Internal error"
+	case CodeTxDecode:
+		return "Tx parse error"
+	case CodeInvalidSequence:
+		return "Invalid sequence"
+	case CodeUnauthorized:
+		return "Unauthorized"
+	case CodeInsufficientFunds:
+		return "Insufficent funds"
+	case CodeUnknownRequest:
+		return "Unknown request"
+	case CodeInvalidAddress:
+		return "Invalid address"
+	case CodeInvalidPubKey:
+		return "Invalid pubkey"
+	case CodeUnknownAddress:
+		return "Unknown address"
+	case CodeInsufficientCoins:
+		return "Insufficient coins"
+	case CodeInvalidCoins:
+		return "Invalid coins"
+	default:
+		return fmt.Sprintf("Unknown code %d", code)
 	}
 }
 
-// Internal Error on Root Codespace
-func ErrInternal(msg string) Error {
-	return NewError(CodespaceRoot, CodeInternal, msg)
-}
+//--------------------------------------------------------------------------------
+// All errors are created via constructors so as to enable us to hijack them
+// and inject stack traces if we really want to.
 
-// Unknown Request Error on Root Codespace
-func ErrUnknownRequest(msg string) Error {
-	return NewError(CodespaceRoot, CodeUnknownRequest, msg)
+// nolint
+func ErrInternal(msg string) sdk.Error {
+	return sdk.NewError(CodespaceSDK, CodeInternal, msg)
 }
-
-// Unknown TxDecode Error on Root Codespace
-func ErrTxDecode(msg string) Error {
-	return NewError(CodespaceRoot, CodeTxDecode, msg)
+func ErrTxDecode(msg string) sdk.Error {
+	return sdk.NewError(CodespaceSDK, CodeTxDecode, msg)
 }
-
-// StdError -- implements Error interface
-type StdError struct {
-	codespace CodespaceType
-	code      CodeType
-	err       cmn.Error
+func ErrInvalidSequence(msg string) sdk.Error {
+	return sdk.NewError(CodespaceSDK, CodeInvalidSequence, msg)
 }
-
-// Implements ABCIError.
-func (err *StdError) Error() string {
-	return fmt.Sprintf("Error{%d:%d,%#v}", err.codespace, err.code, err.err)
+func ErrUnauthorized(msg string) sdk.Error {
+	return sdk.NewError(CodespaceSDK, CodeUnauthorized, msg)
 }
-
-// Implements ABCIError.
-func (err *StdError) ABCICode() ABCICodeType {
-	return ToABCICode(err.codespace, err.code)
+func ErrInsufficientFunds(msg string) sdk.Error {
+	return sdk.NewError(CodespaceSDK, CodeInsufficientFunds, msg)
 }
-
-// Implements Error.
-func (err *StdError) Codespace() CodespaceType {
-	return err.codespace
+func ErrUnknownRequest(msg string) sdk.Error {
+	return sdk.NewError(CodespaceSDK, CodeUnknownRequest, msg)
 }
-
-// Implements Error.
-func (err *StdError) Code() CodeType {
-	return err.code
+func ErrInvalidAddress(msg string) sdk.Error {
+	return sdk.NewError(CodespaceSDK, CodeInvalidAddress, msg)
 }
-
-// Implements ABCIError.
-func (err *StdError) ABCILog() string {
-	return fmt.Sprintf(`=== ABCI Log ===
-Codespace: %v
-Code:      %v
-ABCICode:  %v
-Error:     %#v
-=== /ABCI Log ===
-`, err.codespace, err.code, err.ABCICode(), err.err)
+func ErrUnknownAddress(msg string) sdk.Error {
+	return sdk.NewError(CodespaceSDK, CodeUnknownAddress, msg)
 }
-
-// Add tracing information with msg.
-func (err *StdError) Trace(msg string) Error {
-	return &StdError{
-		codespace: err.codespace,
-		code:      err.code,
-		err:       err.err.Trace(msg),
-	}
+func ErrInvalidPubKey(msg string) sdk.Error {
+	return sdk.NewError(CodespaceSDK, CodeInvalidPubKey, msg)
 }
-
-// Implements Error.
-func (err *StdError) WithDefaultCodespace(cs CodespaceType) Error {
-	codespace := err.codespace
-	if codespace == CodespaceUndefined {
-		codespace = cs
-	}
-	return &StdError{
-		codespace: codespace,
-		code:      err.code,
-		err:       err.err,
-	}
+func ErrInsufficientCoins(msg string) sdk.Error {
+	return sdk.NewError(CodespaceSDK, CodeInsufficientCoins, msg)
 }
-
-func (err *StdError) T() interface{} {
-	return err.err.T()
-}
-
-func (err *StdError) Result() Result {
-	return Result{
-		Code: err.ABCICode(),
-		Log:  err.ABCILog(),
-	}
-}
-
-// QueryResult allows us to return sdk.Error.QueryResult() in query responses
-func (err *StdError) QueryResult() abci.ResponseQuery {
-	return abci.ResponseQuery{
-		Code: uint32(err.ABCICode()),
-		Log:  err.ABCILog(),
-	}
+func ErrInvalidCoins(msg string) sdk.Error {
+	return sdk.NewError(CodespaceSDK, CodeInvalidCoins, msg)
 }
