+++ baseapp/rational_test.go
@@ -1,4 +1,4 @@
-package types
+package baseapp
 
 import (
 	"math/big"
