+++ store/cachekvstore.go
@@ -5,7 +5,7 @@ import (
 	"sort"
 	"sync"
 
-	"github.com/cosmos/cosmos-sdk/baseapp"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	cmn "github.com/tendermint/tmlibs/common"
 )
 
@@ -136,12 +136,12 @@ func (ci *cacheKVStore) ReverseIterator(start, end []byte) Iterator {
 
 // Implements KVStore.
 func (ci *cacheKVStore) SubspaceIterator(prefix []byte) Iterator {
-	return ci.iterator(prefix, baseapp.PrefixEndBytes(prefix), true)
+	return ci.iterator(prefix, sdk.PrefixEndBytes(prefix), true)
 }
 
 // Implements KVStore.
 func (ci *cacheKVStore) ReverseSubspaceIterator(prefix []byte) Iterator {
-	return ci.iterator(prefix, baseapp.PrefixEndBytes(prefix), false)
+	return ci.iterator(prefix, sdk.PrefixEndBytes(prefix), false)
 }
 
 func (ci *cacheKVStore) iterator(start, end []byte, ascending bool) Iterator {
