+++ store/iavlstore_test.go
@@ -3,14 +3,13 @@ package store
 import (
 	"testing"
 
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/assert"
 
 	abci "github.com/tendermint/abci/types"
 	"github.com/tendermint/iavl"
 	cmn "github.com/tendermint/tmlibs/common"
 	dbm "github.com/tendermint/tmlibs/db"
-
-	"github.com/cosmos/cosmos-sdk/baseapp"
 )
 
 var (
@@ -276,19 +275,19 @@ func TestIAVLStoreQuery(t *testing.T) {
 	// set data without commit, doesn't show up
 	iavlStore.Set(k, v)
 	qres := iavlStore.Query(query)
-	assert.Equal(t, uint32(baseapp.CodeOK), qres.Code)
+	assert.Equal(t, uint32(sdk.CodeOK), qres.Code)
 	assert.Nil(t, qres.Value)
 
 	// commit it, but still don't see on old version
 	cid = iavlStore.Commit()
 	qres = iavlStore.Query(query)
-	assert.Equal(t, uint32(baseapp.CodeOK), qres.Code)
+	assert.Equal(t, uint32(sdk.CodeOK), qres.Code)
 	assert.Nil(t, qres.Value)
 
 	// but yes on the new version
 	query.Height = cid.Version
 	qres = iavlStore.Query(query)
-	assert.Equal(t, uint32(baseapp.CodeOK), qres.Code)
+	assert.Equal(t, uint32(sdk.CodeOK), qres.Code)
 	assert.Equal(t, v, qres.Value)
 
 	// modify
@@ -298,22 +297,22 @@ func TestIAVLStoreQuery(t *testing.T) {
 
 	// query will return old values, as height is fixed
 	qres = iavlStore.Query(query)
-	assert.Equal(t, uint32(baseapp.CodeOK), qres.Code)
+	assert.Equal(t, uint32(sdk.CodeOK), qres.Code)
 	assert.Equal(t, v, qres.Value)
 
 	// update to latest in the query and we are happy
 	query.Height = cid.Version
 	qres = iavlStore.Query(query)
-	assert.Equal(t, uint32(baseapp.CodeOK), qres.Code)
+	assert.Equal(t, uint32(sdk.CodeOK), qres.Code)
 	assert.Equal(t, v3, qres.Value)
 	query2 := abci.RequestQuery{Path: "/key", Data: k2, Height: cid.Version}
 	qres = iavlStore.Query(query2)
-	assert.Equal(t, uint32(baseapp.CodeOK), qres.Code)
+	assert.Equal(t, uint32(sdk.CodeOK), qres.Code)
 	assert.Equal(t, v2, qres.Value)
 
 	// default (height 0) will show latest -1
 	query0 := abci.RequestQuery{Path: "/store", Data: k}
 	qres = iavlStore.Query(query0)
-	assert.Equal(t, uint32(baseapp.CodeOK), qres.Code)
+	assert.Equal(t, uint32(sdk.CodeOK), qres.Code)
 	assert.Equal(t, v, qres.Value)
 }
