+++ store/rootmultistore_test.go
@@ -3,12 +3,11 @@ package store
 import (
 	"testing"
 
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/assert"
 	abci "github.com/tendermint/abci/types"
 	dbm "github.com/tendermint/tmlibs/db"
 	"github.com/tendermint/tmlibs/merkle"
-
-	"github.com/cosmos/cosmos-sdk/baseapp"
 )
 
 const useDebugDB = false
@@ -128,34 +127,34 @@ func TestMultiStoreQuery(t *testing.T) {
 	// Test bad path.
 	query := abci.RequestQuery{Path: "/key", Data: k, Height: ver}
 	qres := multi.Query(query)
-	assert.Equal(t, baseapp.ToABCICode(baseapp.CodespaceRoot, baseapp.CodeUnknownRequest), baseapp.ABCICodeType(qres.Code))
+	assert.Equal(t, sdk.ToABCICode(sdk.CodespaceRoot, sdk.CodeUnknownRequest), sdk.ABCICodeType(qres.Code))
 
 	query.Path = "h897fy32890rf63296r92"
 	qres = multi.Query(query)
-	assert.Equal(t, baseapp.ToABCICode(baseapp.CodespaceRoot, baseapp.CodeUnknownRequest), baseapp.ABCICodeType(qres.Code))
+	assert.Equal(t, sdk.ToABCICode(sdk.CodespaceRoot, sdk.CodeUnknownRequest), sdk.ABCICodeType(qres.Code))
 
 	// Test invalid store name.
 	query.Path = "/garbage/key"
 	qres = multi.Query(query)
-	assert.Equal(t, baseapp.ToABCICode(baseapp.CodespaceRoot, baseapp.CodeUnknownRequest), baseapp.ABCICodeType(qres.Code))
+	assert.Equal(t, sdk.ToABCICode(sdk.CodespaceRoot, sdk.CodeUnknownRequest), sdk.ABCICodeType(qres.Code))
 
 	// Test valid query with data.
 	query.Path = "/store1/key"
 	qres = multi.Query(query)
-	assert.Equal(t, baseapp.ToABCICode(baseapp.CodespaceRoot, baseapp.CodeOK), baseapp.ABCICodeType(qres.Code))
+	assert.Equal(t, sdk.ToABCICode(sdk.CodespaceRoot, sdk.CodeOK), sdk.ABCICodeType(qres.Code))
 	assert.Equal(t, v, qres.Value)
 
 	// Test valid but empty query.
 	query.Path = "/store2/key"
 	query.Prove = true
 	qres = multi.Query(query)
-	assert.Equal(t, baseapp.ToABCICode(baseapp.CodespaceRoot, baseapp.CodeOK), baseapp.ABCICodeType(qres.Code))
+	assert.Equal(t, sdk.ToABCICode(sdk.CodespaceRoot, sdk.CodeOK), sdk.ABCICodeType(qres.Code))
 	assert.Nil(t, qres.Value)
 
 	// Test store2 data.
 	query.Data = k2
 	qres = multi.Query(query)
-	assert.Equal(t, baseapp.ToABCICode(baseapp.CodespaceRoot, baseapp.CodeOK), baseapp.ABCICodeType(qres.Code))
+	assert.Equal(t, sdk.ToABCICode(sdk.CodespaceRoot, sdk.CodeOK), sdk.ABCICodeType(qres.Code))
 	assert.Equal(t, v2, qres.Value)
 }
