+++ types/codespacer.go
@@ -1,4 +1,4 @@
-package baseapp
+package sdk
 
 //----------------------------------------
 // Codespacer
