+++ types/data/TestBasic.db/CURRENT
@@ -1,0 +1,1 @@
+MANIFEST-000011
