+++ types/errors.go
@@ -1,98 +1,173 @@
-package types
+package sdk
 
 import (
 	"fmt"
 
-	"github.com/cosmos/cosmos-sdk/baseapp"
+	abci "github.com/tendermint/abci/types"
+	cmn "github.com/tendermint/tmlibs/common"
 )
 
-// SDK error codes
+// ABCICodeType - combined codetype / codespace
+type ABCICodeType uint32
+
+// CodeType - code identifier within codespace
+type CodeType uint16
+
+// CodespaceType - codespace identifier
+type CodespaceType uint16
+
+// root error codes
 const (
+
 	// ABCI error codes
-	ABCICodeOK baseapp.ABCICodeType = 0
+	ABCICodeOK ABCICodeType = 0
 
-	// Base error codes
-	CodeOK                baseapp.CodeType = 0
-	CodeInternal          baseapp.CodeType = 1
-	CodeTxDecode          baseapp.CodeType = 2
-	CodeInvalidSequence   baseapp.CodeType = 3
-	CodeUnauthorized      baseapp.CodeType = 4
-	CodeInsufficientFunds baseapp.CodeType = 5
-	CodeUnknownRequest    baseapp.CodeType = 6
-	CodeInvalidAddress    baseapp.CodeType = 7
-	CodeInvalidPubKey     baseapp.CodeType = 8
-	CodeUnknownAddress    baseapp.CodeType = 9
-	CodeInsufficientCoins baseapp.CodeType = 10
-	CodeInvalidCoins      baseapp.CodeType = 11
+	// Root error codes
+	CodeOK             CodeType = 0
+	CodeInternal       CodeType = 1
+	CodeUnknownRequest CodeType = 2
+	CodeTxDecode       CodeType = 3
 
-	CodespaceSDK baseapp.CodespaceType = 2
+	// CodespaceRoot is a codespace for error codes in this file only.
+	// Notice that 0 is an "unset" codespace, which can be overridden with
+	// Error.WithDefaultCodespace().
+	CodespaceUndefined CodespaceType = 0
+	CodespaceRoot      CodespaceType = 1
+
+	// Maximum reservable codespace (2^16 - 1)
+	MaximumCodespace CodespaceType = 65535
 )
 
-// NOTE: Don't stringer this, we'll put better messages in later.
-func CodeToDefaultMsg(code baseapp.CodeType) string {
-	switch code {
-	case CodeInternal:
-		return "Internal error"
-	case CodeTxDecode:
-		return "Tx parse error"
-	case CodeInvalidSequence:
-		return "Invalid sequence"
-	case CodeUnauthorized:
-		return "Unauthorized"
-	case CodeInsufficientFunds:
-		return "Insufficent funds"
-	case CodeUnknownRequest:
-		return "Unknown request"
-	case CodeInvalidAddress:
-		return "Invalid address"
-	case CodeInvalidPubKey:
-		return "Invalid pubkey"
-	case CodeUnknownAddress:
-		return "Unknown address"
-	case CodeInsufficientCoins:
-		return "Insufficient coins"
-	case CodeInvalidCoins:
-		return "Invalid coins"
-	default:
-		return fmt.Sprintf("Unknown code %d", code)
+// Error type
+type Error interface {
+	Error() string
+	Code() CodeType
+	Codespace() CodespaceType
+	ABCILog() string
+	ABCICode() ABCICodeType
+	WithDefaultCodespace(codespace CodespaceType) Error
+	Trace(msg string) Error
+	T() interface{}
+	Result() Result
+	QueryResult() abci.ResponseQuery
+}
+
+// get the abci code from the local code and codespace
+func ToABCICode(space CodespaceType, code CodeType) ABCICodeType {
+	// TODO: Make Tendermint more aware of codespaces.
+	if space == CodespaceRoot && code == CodeOK {
+		return ABCICodeOK
+	}
+	return ABCICodeType((uint32(space) << 16) | uint32(code))
+}
+
+// IsOK - is everything okay?
+func (code ABCICodeType) IsOK() bool {
+	if code == ABCICodeOK {
+		return true
+	}
+	return false
+}
+
+// NewError - create an error
+func NewError(codespace CodespaceType, code CodeType, msg string) Error {
+	return &StdError{
+		codespace: codespace,
+		code:      code,
+		err:       cmn.NewErrorWithT(code, msg),
 	}
 }
 
-//--------------------------------------------------------------------------------
-// All errors are created via constructors so as to enable us to hijack them
-// and inject stack traces if we really want to.
+// Internal Error on Root Codespace
+func ErrInternal(msg string) Error {
+	return NewError(CodespaceRoot, CodeInternal, msg)
+}
 
-// nolint
-func ErrInternal(msg string) baseapp.Error {
-	return baseapp.NewError(CodespaceSDK, CodeInternal, msg)
+// Unknown Request Error on Root Codespace
+func ErrUnknownRequest(msg string) Error {
+	return NewError(CodespaceRoot, CodeUnknownRequest, msg)
 }
-func ErrTxDecode(msg string) baseapp.Error {
-	return baseapp.NewError(CodespaceSDK, CodeTxDecode, msg)
+
+// Unknown TxDecode Error on Root Codespace
+func ErrTxDecode(msg string) Error {
+	return NewError(CodespaceRoot, CodeTxDecode, msg)
 }
-func ErrInvalidSequence(msg string) baseapp.Error {
-	return baseapp.NewError(CodespaceSDK, CodeInvalidSequence, msg)
+
+// StdError -- implements Error interface
+type StdError struct {
+	codespace CodespaceType
+	code      CodeType
+	err       cmn.Error
 }
-func ErrUnauthorized(msg string) baseapp.Error {
-	return baseapp.NewError(CodespaceSDK, CodeUnauthorized, msg)
+
+// Implements ABCIError.
+func (err *StdError) Error() string {
+	return fmt.Sprintf("Error{%d:%d,%#v}", err.codespace, err.code, err.err)
 }
-func ErrInsufficientFunds(msg string) baseapp.Error {
-	return baseapp.NewError(CodespaceSDK, CodeInsufficientFunds, msg)
+
+// Implements ABCIError.
+func (err *StdError) ABCICode() ABCICodeType {
+	return ToABCICode(err.codespace, err.code)
 }
-func ErrUnknownRequest(msg string) baseapp.Error {
-	return baseapp.NewError(CodespaceSDK, CodeUnknownRequest, msg)
+
+// Implements Error.
+func (err *StdError) Codespace() CodespaceType {
+	return err.codespace
 }
-func ErrInvalidAddress(msg string) baseapp.Error {
-	return baseapp.NewError(CodespaceSDK, CodeInvalidAddress, msg)
+
+// Implements Error.
+func (err *StdError) Code() CodeType {
+	return err.code
 }
-func ErrUnknownAddress(msg string) baseapp.Error {
-	return baseapp.NewError(CodespaceSDK, CodeUnknownAddress, msg)
+
+// Implements ABCIError.
+func (err *StdError) ABCILog() string {
+	return fmt.Sprintf(`=== ABCI Log ===
+Codespace: %v
+Code:      %v
+ABCICode:  %v
+Error:     %#v
+=== /ABCI Log ===
+`, err.codespace, err.code, err.ABCICode(), err.err)
 }
-func ErrInvalidPubKey(msg string) baseapp.Error {
-	return baseapp.NewError(CodespaceSDK, CodeInvalidPubKey, msg)
+
+// Add tracing information with msg.
+func (err *StdError) Trace(msg string) Error {
+	return &StdError{
+		codespace: err.codespace,
+		code:      err.code,
+		err:       err.err.Trace(msg),
+	}
 }
-func ErrInsufficientCoins(msg string) baseapp.Error {
-	return baseapp.NewError(CodespaceSDK, CodeInsufficientCoins, msg)
+
+// Implements Error.
+func (err *StdError) WithDefaultCodespace(cs CodespaceType) Error {
+	codespace := err.codespace
+	if codespace == CodespaceUndefined {
+		codespace = cs
+	}
+	return &StdError{
+		codespace: codespace,
+		code:      err.code,
+		err:       err.err,
+	}
 }
-func ErrInvalidCoins(msg string) baseapp.Error {
-	return baseapp.NewError(CodespaceSDK, CodeInvalidCoins, msg)
+
+func (err *StdError) T() interface{} {
+	return err.err.T()
+}
+
+func (err *StdError) Result() Result {
+	return Result{
+		Code: err.ABCICode(),
+		Log:  err.ABCILog(),
+	}
+}
+
+// QueryResult allows us to return sdk.Error.QueryResult() in query responses
+func (err *StdError) QueryResult() abci.ResponseQuery {
+	return abci.ResponseQuery{
+		Code: uint32(err.ABCICode()),
+		Log:  err.ABCILog(),
+	}
 }
